requirements.txt
flask==2.3.3
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
sseclient-py==1.8.0
Flask-Cors==4.0.0
gevent==24.2.1
